changegreedy & driver for command line input
Modified existing functions we used from project 1, and added a few new ones.  runAlgorithms.py is the file that can be run from the command-line with a .txt input file. 
i.e. ./python runAlgorithms.py "inputFilename.txt"
This is working on Windows, but I haven't run it on Linux yet.  Once the other algorithms are up we'll have to make a few changes to processTestData.py