attempt at changeslow
This isn't quite there, but I think it's really close.  I am having trouble using recursion with returning a tuple.  If I only return the total number of coins needed, this works exactly like it should, and from what I can tell using the print statement it tries all possible combinations of coins.  I'm hoping you guys have a little more experience with Python, and might be able to see an easy way to include the array of each denomination coin necessary in addition to the total number of coins.